Add files via upload
LM7805

 

Features

• Output Current up to 1.5 A

• Available in Fixed 5-V, 12-V, and 15-V Options

• Output Voltage Tolerances of ±2% at TJ = 25°C (LM340A)

• Line Regulation of 0.01% / V of at 1-A Load (LM340A)

 • Load Regulation of 0.3% / A (LM340A)

• Internal Thermal Overload, Short-Circuit and SOA Protection

• Available in Space-Saving SOT-223 Package

• Output Capacitance Not Required for Stability 2 Applications

• Industrial Power Supplies

• SMPS Post Regulation

• HVAC Systems

• AC Inventors

• Test and Measurement Equipment

• Brushed and Brushless DC Motor Drivers

• Solar Energy String Invertors



 Description

The LM340 and LM7805 Family monolithic 3-terminal positive voltage regulators employ internal currentlimiting, thermal shutdown and safe-area compensation, making them essentially indestructible. If adequate heat sinking is provided, they can deliver over 1.5-A output current. They are intended as fixed voltage regulators in a wide range of applications including local (on-card) regulation for elimination of noise and distribution problems associated with single-point regulation. In addition to use as fixed voltage regulators, these devices can be used with external components to obtain adjustable output voltages and currents. Considerable effort was expended to make the entire series of regulators easy to use and minimize the number of external components. It is not necessary to bypass the output, although this does improve transient response. Input bypassing is needed only if the regulator is located far from the filter capacitor of the power supply. LM7805 is also available in a higher accuracy and better performance version (LM340A). R